[RENAME]/#4 PackingList를 List로 변경
